PolAssessment/AnprFrontEnd/requirements.txt
Flask==3.0.1
Flask-Minify==0.42
Frozen-Flask==1.0.1
Jinja2==3.1.3
python-dotenv==1.0.1
